chore: moved OpenAPI generated Java BRC client code to separate package
