+++ src/main/java/net/atos/client/zgw/brc/BRCClientService.java
@@ -16,8 +16,8 @@ import jakarta.inject.Inject;
 
 import org.eclipse.microprofile.rest.client.inject.RestClient;
 
-import net.atos.client.zgw.brc.model.Besluit;
-import net.atos.client.zgw.brc.model.BesluitInformatieObject;
+import net.atos.client.zgw.brc.model.generated.Besluit;
+import net.atos.client.zgw.brc.model.generated.BesluitInformatieObject;
 import net.atos.client.zgw.brc.model.BesluitenListParameters;
 import net.atos.client.zgw.shared.model.Results;
 import net.atos.client.zgw.shared.model.audit.AuditTrailRegel;
