+++ src/main/java/net/atos/zac/app/zaken/ZakenRESTService.java
@@ -60,8 +60,8 @@ import net.atos.client.or.object.model.ORObject;
 import net.atos.client.vrl.VRLClientService;
 import net.atos.client.vrl.model.CommunicatieKanaal;
 import net.atos.client.zgw.brc.BRCClientService;
-import net.atos.client.zgw.brc.model.Besluit;
-import net.atos.client.zgw.brc.model.BesluitInformatieObject;
+import net.atos.client.zgw.brc.model.generated.Besluit;
+import net.atos.client.zgw.brc.model.generated.BesluitInformatieObject;
 import net.atos.client.zgw.drc.DRCClientService;
 import net.atos.client.zgw.drc.model.generated.EnkelvoudigInformatieObject;
 import net.atos.client.zgw.shared.ZGWApiService;
