+++ src/main/java/net/atos/zac/app/zaken/converter/RESTBesluitConverter.java
@@ -12,7 +12,7 @@ import java.util.stream.Collectors;
 import jakarta.inject.Inject;
 
 import net.atos.client.zgw.brc.BRCClientService;
-import net.atos.client.zgw.brc.model.Besluit;
+import net.atos.client.zgw.brc.model.generated.Besluit;
 import net.atos.client.zgw.drc.DRCClientService;
 import net.atos.client.zgw.drc.model.generated.EnkelvoudigInformatieObject;
 import net.atos.client.zgw.zrc.model.Zaak;
