+++ src/main/java/net/atos/zac/websocket/event/ScreenEventType.java
@@ -17,7 +17,7 @@ import java.util.UUID;
 
 import org.flowable.task.api.TaskInfo;
 
-import net.atos.client.zgw.brc.model.Besluit;
+import net.atos.client.zgw.brc.model.generated.Besluit;
 import net.atos.client.zgw.drc.model.generated.EnkelvoudigInformatieObject;
 import net.atos.client.zgw.zrc.model.Zaak;
 import net.atos.zac.event.Opcode;
